+++ arrayfire/src/api/cpp/approx.cpp
@@ -27,7 +27,7 @@ array approx2(const array &zi, const array &xo, const array &yo,
 }
 
 array approx1(const array &yi, const array &xo, const int xdim,
-              const double xi_beg, const double xi_step,
+              const float xi_beg, const float xi_step,
               const interpType method, const float offGrid) {
     af_array yo = 0;
     AF_THROW(af_approx1_uniform(&yo, yi.get(), xo.get(), xdim, xi_beg, xi_step,
@@ -36,8 +36,8 @@ array approx1(const array &yi, const array &xo, const int xdim,
 }
 
 array approx2(const array &zi, const array &xo, const int xdim,
-              const double xi_beg, const double xi_step, const array &yo,
-              const int ydim, const double yi_beg, const double yi_step,
+              const float xi_beg, const float xi_step, const array &yo,
+              const int ydim, const float yi_beg, const float yi_step,
               const interpType method, const float offGrid) {
     af_array zo = 0;
     AF_THROW(af_approx2_uniform(&zo, zi.get(), xo.get(), xdim, xi_beg, xi_step,
